backend/requirements.txt
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
django-vite==3.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
pip==25.0.1
pipreqs==0.4.13
psycopg-binary==3.2.6
python-decouple>=3.8
numpy>=2.2.5
psycopg2-binary>=2.9.10
dj-rest-auth>=7.0.1
django-allauth>=65.9.0
cryptography>=45.0.4
